requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-Session==0.5.0
mysql-connector-python==8.1.0
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.0
pygogo==1.3.0
python-dotenv==1.0.0
XlsxWriter==3.1.4
